requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-seed==0.1.9
djangorestframework==3.9.1
Faker==1.0.2
gunicorn==19.9.0
mysqlclient==1.3.14
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
text-unidecode==1.2
whitenoise==4.1.2
geopy==1.18.1
django-session-security==2.6.5
django-excel== 0.0.1
sendgrid==6.1.0
django-filter==2.2.0
fcm-django==0.3.4
mysql-connector-python==8.0.19


asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-excel==0.0.1
django-filter==2.2.0
django-heroku==0.3.1
django-seed==0.1.9
django-session-security==2.6.5
djangorestframework==3.9.1
dnspython==1.16.0
Faker==1.0.2
fcm-django==0.3.4
geographiclib==1.50
geopy==1.18.1
gunicorn==19.9.0
html5lib==1.1
idna==2.10
lml==0.0.9
mysql-connector-python==8.0.19
mysqlclient==1.3.14
Pillow==5.4.1
protobuf==3.6.1
psycopg2==2.7.7
pycparser==2.20
pyexcel==0.6.4
pyexcel-io==0.5.20
pyfcm==1.4.7
PyMySQL==0.10.1
Pyphen==0.9.5
python-dateutil==2.7.5
python-http-client==3.3.1
pytz==2018.9
requests==2.24.0
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
text-unidecode==1.2
texttable==1.6.3
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
whitenoise==4.1.2
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-excel==0.0.1
django-filter==2.2.0
django-heroku==0.3.1
django-seed==0.1.9
django-session-security==2.6.5
djangorestframework==3.9.1
dnspython==1.16.0
Faker==1.0.2
fcm-django==0.3.4
geographiclib==1.50
geopy==1.18.1
gunicorn==19.9.0
html5lib==1.1
idna==2.10
lml==0.0.9
mysql-connector-python==8.0.19
mysqlclient==1.3.14
Pillow==5.4.1
protobuf==3.6.1
psycopg2==2.7.7
pycparser==2.20
pyexcel==0.6.4
pyexcel-io==0.5.20
pyfcm==1.4.7
PyMySQL==0.10.1
Pyphen==0.9.5
python-dateutil==2.7.5
python-http-client==3.3.1
pytz==2018.9
requests==2.24.0
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
text-unidecode==1.2
texttable==1.6.3
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
whitenoise==4.1.2
